restaurant_pipeline/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
rapidfuzz>=3.5.0
geopy>=2.4.0
haversine>=2.8.0
transliterate>=1.10.2
structlog>=23.2.0
tqdm>=4.66.0
click>=8.1.0
python-dotenv>=1.0.0
